Group tests into one file
