+++ kong/tests/common.py
@@ -17,17 +17,3 @@ instance_1 = {'kong_status_url': STATUS_URL, 'tags': ['first_instance']}
 instance_2 = {'kong_status_url': STATUS_URL, 'tags': ['second_instance']}
 
 CONFIG_STUBS = [instance_1, instance_2]
-
-BAD_CONFIG = {'kong_status_url': 'http://localhost:1111/status/'}
-
-EXPECTED_GAUGES = [
-    'kong.total_requests',
-    'kong.connections_active',
-    'kong.connections_waiting',
-    'kong.connections_reading',
-    'kong.connections_accepted',
-    'kong.connections_writing',
-    'kong.connections_handled',
-]
-
-DATABASES = ['reachable']
